% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{computeICC_random}
\alias{computeICC_random}
\title{Calculate random effects LME variances into ICC - BETA VERSION}
\usage{
computeICC_random(lme_model, subjects = "subjects")
}
\arguments{
\item{subjects}{The specific variable designated as the raters for ICC calculation.}

\item{model_fit}{The model object from a lme4.}
}
\value{
A vector of ICCs from the specified lme4 model
}
\description{
Given a random LME model from LME4, first identify the model, then calculate ICC from variances
}
