% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varde-package.R
\docType{package}
\name{varde-package}
\alias{varde}
\alias{varde-package}
\title{varde: Variance Decomposition Functions}
\description{
Functions for doing variance decomposition, e.g., for conducting generalizability studies or estimating intraclass correlations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jmgirard/varde}
  \item Report bugs at \url{https://github.com/jmgirard/varde/issues}
}

}
\author{
\strong{Maintainer}: Jeffrey Girard \email{me@jmgirard.com}

}
\keyword{internal}
